+++ src/stream_wrapper.hh
@@ -107,7 +107,7 @@ namespace crouton {
         void read(ssize_t nread, const uv_buf_t* uvbuf) {
             assert(nread != 0);
             if (nread > 0) {
-                assert(nread <= Buffer::kCapacity);
+                assert(size_t(nread) <= Buffer::kCapacity);
                 assert(uvbuf->base == _readingBuf->data);
                 _readingBuf->length = uint32_t(nread);
                 _readingBuf->used = 0;
